pyproject.toml
[project]
name = "udd-skill-bundle"
version = "0.1.0"
description = "AX Universal Data Downloader — skill bundle"
requires-python = ">=3.10"
authors = [{name = "gibbs hwang"}]
# Runtime deps the skill helper scripts (precheck, scope, scaffold,
# auth_flow, refactor, validate_loop, ...) need on the *host* Python.
# These are NOT the deps the generated user project gets — those live
# in templates/requirements.txt.tmpl. Skill scripts call
# ``lib.llm_call.ask`` which imports anthropic / google-generativeai /
# openai lazily, so all three are listed as optional extras (install at
# least one). pyyaml is unconditional — every script reads config.yaml.
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.34"]
gemini = ["google-generativeai>=0.7"]
openai = ["openai>=1.40"]
dev = [
  "pytest>=8.0",
  "pytest-mock>=3.12",
  "pyyaml>=6.0",
  "respx>=0.22",
  "pandas>=2.0",
  "openpyxl>=3.1",
  "keyring>=24.0",
  "requests>=2.31",
  "anthropic>=0.34",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
